More renaming (debated in rocket chat); works for WASM build, but not yet for unit tests.
