+++ src/storage/data_object_type_registry.rs
