+++ src/storage/mod.rs
@@ -1,6 +1,6 @@
 #![cfg_attr(not(feature = "std"), no_std)]
 
-pub mod type_registry;
+pub mod data_object_type_registry;
 
 mod mock;
 mod tests;
